dev-requirements.txt
-r requirements.txt

black == 20.8b1
bumpversion == 0.6.0
coverage == 5.3
factory_boy == 3.1.0
flake8 == 3.8.4
pyaml == 20.4.0
pytest == 6.1.2
pytest-django == 4.1.0
